requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
Markdown==3.7
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
